integracion-2/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
transbank-sdk==3.8.0
